Ensure that <[a-z] cannot be in stylesheets, as it looks like html
In foreign content (svg) the rules are different and anything that
looks like a tag can make one, where in normal html it is just
</style. Block both just in case.

Also disallow <! as a paranoid precaution against comments
and CDATA sections.

While this doesn't fully depend on I8dfa5c7b67caa31ba6ea
it is reccomended that it is deployed with it.

Bug: T381617
Change-Id: I2cb4d046774dd8c8a446f1d072c970c3bc05f95d