Fix adding warnings to ParserOutput
ParserOutput::addWarningMsg() accepts the params as variadic
arguments, not as an array (should have been `...$error['params']`).

This is easy to get wrong, so instead, let's use a newer style that
passes the message with params as a whole, instead of separating them.
This replaces the use of deprecated StatusValue::getErrors() and
soon-to-be-deprecated ParserOutput::addWarningMsg().

Bug: T396768
Change-Id: I61ac04459e5e4ee9ff6c0ed1d774858d01ad959a
(cherry picked from commit 904c65118cb80bc926fb6aff3f36f3f4729e6724)